backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
alembic==1.12.1 